Moved checkoutPage test to CheckoutPageTest class
